requirements.txt
appdirs==1.4.3
attrs==20.3.0
boto3==1.17.22
botocore==1.20.22
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
coloredlogs==15.0
contextlib2==0.6.0
cryptography==3.4.6
distlib==0.3.0
distro==1.4.0
docopt==0.6.2
html5lib==1.0.1
humanfriendly==9.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
jmespath==0.10.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
py==1.10.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.6
pytest==6.2.2
python-dateutil==2.8.1
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.4
sewer==0.8.4
six==1.14.0
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
